rename search query
